server/requirements.txt
fastapi==0.110.0
pyairtable==2.3.1
python-dotenv==1.0.1
uvicorn==0.27.1
pydantic==2.6.3
python-multipart==0.0.9
